Move suppliers to textFixtures
