+++ infrastructure/kzg/src/testFixtures/java/tech/pegasys/teku/kzg/propertytest/suppliers/KZGCommitmentSupplier.java
@@ -11,15 +11,15 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.kzg.suppliers;
+package tech.pegasys.teku.kzg.propertytest.suppliers;
 
-import tech.pegasys.teku.kzg.KZGProof;
+import tech.pegasys.teku.kzg.KZGCommitment;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.propertytest.suppliers.DataStructureUtilSupplier;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
-public class KZGProofSupplier extends DataStructureUtilSupplier<KZGProof> {
-  public KZGProofSupplier() {
-    super(DataStructureUtil::randomKZGProof, SpecMilestone.EIP4844);
+public class KZGCommitmentSupplier extends DataStructureUtilSupplier<KZGCommitment> {
+  public KZGCommitmentSupplier() {
+    super(DataStructureUtil::randomKZGCommitment, SpecMilestone.EIP4844);
   }
 }
